Added integrationtest which proves that delete journey fail of journey has tag
